Fix MdaExtractors + tests
